Add support for Cloud Quota Management API
To work with "limits" and "quotas", we now need to handle
the Cloud Quota Management API instead of the Cloud Management API.

QuotaRegionalClient is a Proxy of the HTTP client. It resolves
the URL of the service in the requested region and adds X-Auth-Token to
the headers.

The functionality for issuing and caching the X-Auth-Token, as well as
resolving the service URL is encapsulated in the IdentityManager.